consumer/requirements.txt
kafka-python==2.0.2
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
eventlet==0.36.1
